Cleaner code
